iommu: Move IOMMU mapping out of Ion directory
The IOMMU mapping APIs used to be tied to Ion but there
is less of a dependency these days. Move the file out of
the Ion directory and to somewhere more generic.
