+++ drivers/staging/android/ion/msm/Makefile
@@ -1,4 +1,4 @@
-obj-y += msm_ion.o ion_iommu_map.o secure_buffer.o
+obj-y += msm_ion.o secure_buffer.o
 ifdef CONFIG_COMPAT
 obj-y += compat_msm_ion.o
 endif
