+++ include/linux/msm_iommu_domains.h
@@ -17,6 +17,7 @@
 #include <linux/mutex.h>
 #include <linux/genalloc.h>
 #include <linux/rbtree.h>
+#include <linux/msm_ion.h>
 
 #define MSM_IOMMU_DOMAIN_SECURE	0x1
 
@@ -87,6 +88,45 @@ struct msm_iova_layout {
 };
 
 #if defined(CONFIG_MSM_IOMMU)
+/**
+ * ion_map_iommu - map the given handle into an iommu
+ *
+ * @client - client who allocated the handle
+ * @handle - handle to map
+ * @domain_num - domain number to map to
+ * @partition_num - partition number to allocate iova from
+ * @align - alignment for the iova
+ * @iova_length - length of iova to map. If the iova length is
+ *		greater than the handle length, the remaining
+ *		address space will be mapped to a dummy buffer.
+ * @iova - pointer to store the iova address
+ * @buffer_size - pointer to store the size of the buffer
+ * @flags - flags for options to map
+ * @iommu_flags - flags specific to the iommu.
+ *
+ * Maps the handle into the iova space specified via domain number. Iova
+ * will be allocated from the partition specified via partition_num.
+ * Returns 0 on success, negative value on error.
+ */
+int ion_map_iommu(struct ion_client *client, struct ion_handle *handle,
+			int domain_num, int partition_num, unsigned long align,
+			unsigned long iova_length, ion_phys_addr_t *iova,
+			unsigned long *buffer_size,
+			unsigned long flags, unsigned long iommu_flags);
+
+/**
+ * ion_unmap_iommu - unmap the handle from an iommu
+ *
+ * @client - client who allocated the handle
+ * @handle - handle to unmap
+ * @domain_num - domain to unmap from
+ * @partition_num - partition to unmap from
+ *
+ * Decrement the reference count on the iommu mapping. If the count is
+ * 0, the mapping will be removed from the iommu.
+ */
+void ion_unmap_iommu(struct ion_client *client, struct ion_handle *handle,
+			int domain_num, int partition_num);
 
 extern void msm_iommu_set_client_name(struct iommu_domain *domain,
 				      char const *name);
@@ -136,6 +176,24 @@ extern int msm_register_domain(struct msm_iova_layout *layout);
 extern int msm_unregister_domain(struct iommu_domain *domain);
 
 #else
+static inline int ion_map_iommu(struct ion_client *client,
+			struct ion_handle *handle, int domain_num,
+			int partition_num, unsigned long align,
+			unsigned long iova_length, ion_phys_addr_t *iova,
+			unsigned long *buffer_size,
+			unsigned long flags,
+			unsigned long iommu_flags)
+{
+	return -ENODEV;
+}
+
+static inline void ion_unmap_iommu(struct ion_client *client,
+			struct ion_handle *handle, int domain_num,
+			int partition_num)
+{
+	return;
+}
+
 static inline void msm_iommu_set_client_name(struct iommu_domain *domain,
 					     char const *name)
 {
